Updated D&I overview page with A11y and Digital4Good initiatives
The following paragraphs were added in.

paragraph 2:
In 2019, the Digital4Good programme was established to empower underprivileged children/youths and seniors by increasing their digital literacy. This is achieved by inculcating skillsets such as programming, problem solving and critical thinking for children/youth and the use of digital tools for seniors. 

paragrah 4:
In 2022, GovTech’s Accessibility Enabling team was formed to understand and design inclusive government services based on the unique perspectives and usage patterns of persons with disabilities (PwDs) and seniors. The team complements agile development methods and promotes the use of automated accessibility testing in the product development process. This is achieved through the provision of consultancy services, training sessions, distribution of guides and playbooks, and outreach activities.